code split; use class vars instead of global vars
