+++ mopidy_ytmusic/playback.py
@@ -1,0 +1,94 @@
+from urllib.parse import parse_qs
+from mopidy import backend, httpclient
+from mopidy_ytmusic import logger
+from youtube_dl import YoutubeDL
+
+class YoutubeMusicPlaybackProvider(backend.PlaybackProvider):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.last_id = None
+        self.Youtube_Player_URL = None
+        self.YoutubeDL = YoutubeDL({
+            "proxy": httpclient.format_proxy(self.backend.config["proxy"]),
+            "nocheckcertificate": True,
+        })
+        self.YoutubeIE = self.YoutubeDL.get_info_extractor('Youtube')
+
+    def translate_uri(self, uri):
+        logger.info('YTMusic PlaybackProvider.translate_uri "%s"', uri)
+
+        if "ytmusic:track:" not in uri:
+            return None
+
+        try:
+            bId = uri.split(":")[2]
+            self.last_id = bId
+            return self._get_track(bId)
+        except Exception as e:
+            logger.error('translate_uri error "%s"', str(e))
+            return None
+
+    def _get_track(self,bId):
+        streams = self.backend.api.get_streaming_data(bId)
+        # Try to find the highest quality stream.  We want "AUDIO_QUALITY_HIGH", barring
+        # that we find the highest bitrate audio/mp4 stream, after that we sort through the
+        # garbage.
+        playstr = None
+        url = None
+        if 'adaptiveFormats' in streams:
+            bitrate = 0
+            crap = {}
+            worse = {}
+            for stream in streams['adaptiveFormats']:
+                if 'audioQuality' in stream and stream['audioQuality'] == 'AUDIO_QUALITY_HIGH':
+                    playstr = stream
+                    break
+                if stream['mimeType'].startswith('audio/mp4') and stream['averageBitrate'] > bitrate:
+                    bitrate = stream['averageBitrate']
+                    playstr = stream
+                elif stream['mimeType'].startswith('audio'):
+                    crap[stream['averageBitrate']] = stream
+                else:
+                    worse[stream['averageBitrate']] = stream
+            if playstr is None:
+                # sigh.
+                if len(crap):
+                    playstr=crap[sorted(list(crap.keys()))[-1]]
+                    if 'audioQuality' not in playstr:
+                        playstr['audioQuality'] = 'AUDIO_QUALITY_GARBAGE'
+                elif len(worse):
+                    playstr=worse[sorted(list(worse.keys()))[-1]]
+                    if 'audioQuality' not in playstr:
+                        playstr['audioQuality'] = 'AUDIO_QUALITY_FECES'
+        elif 'formats' in streams:
+            # Great, we're really left with the dregs of quality.
+            playstr = streams['formats'][0]
+            if 'audioQuality' not in playstr:
+                playstr['audioQuality'] = 'AUDIO_QUALITY_404'
+            if 'url' in playstr:
+                logger.info('Found %s stream with %d ABR for %s',playstr['audioQuality'],playstr['averageBitrate'],bId)
+                url = playstr['url']
+            else:
+                logger.error('No url for %s.',bId)
+        else:
+            logger.error('No streams found for %s. Falling back to youtube-dl.',bId)
+        if playstr is not None:
+            # Use Youtube-DL's Info Extractor to decode the signature.
+            if 'signatureCipher' in playstr:
+                sc = parse_qs(playstr['signatureCipher'])
+                sig = self.YoutubeIE._decrypt_signature(
+                    sc['s'][0],
+                    bId,
+                    self.Youtube_Player_URL,
+                )
+                url = sc['url'][0] + '&sig=' + sig + '&ratebypass=yes'
+            elif 'url' in playstr:
+                url = playstr['url']
+            else:
+                logger.error("Unable to get URL from stream for %s",bId)
+                return(None)
+            logger.info('Found %s stream with %d ABR for %s',playstr['audioQuality'],playstr['averageBitrate'],bId)
+        if url is not None:
+            # Return the decoded youtube url to mopidy for playback.
+            return(url)
+        return None
